Feature adds

+++ www/js/group/singleGroup/singleGroup.js
@@ -1,6 +1,6 @@
 'use strict';
 
-angular.module('main').controller('SingleGroupCtrl',['$scope', '$state', '$stateParams', '$log', '$ionicNavBarDelegate', 'CameraFactory', '$cordovaCamera', 'loggedInUser', 'GroupFactory', 'GroupService',function ($scope, $state, $stateParams, $log, $ionicNavBarDelegate, CameraFactory, $cordovaCamera, loggedInUser, GroupFactory, GroupService) {
+angular.module('main').controller('SingleGroupCtrl',['$scope', '$state', '$stateParams', '$log', '$ionicNavBarDelegate', 'CameraFactory', '$cordovaCamera', 'loggedInUser', 'GroupFactory', 'GroupService', '$ionicHistory', '$ionicPopup', function ($scope, $state, $stateParams, $log, $ionicNavBarDelegate, CameraFactory, $cordovaCamera, loggedInUser, GroupFactory, GroupService, $ionicHistory, $ionicPopup) {
   var unbind;
   $scope.$on("$ionicView.loaded", function () {
     if($scope.group){
@@ -11,7 +11,7 @@ angular.module('main').controller('SingleGroupCtrl',['$scope', '$state', '$state
     .then(function(group){
       return group.$bindTo($scope, 'group');
     })
-    .then(function(ub){
+    .then(function(ub) {
       unbind = ub;
       $scope.isAdmin = $scope.group.members[loggedInUser.uid];
       $scope.newGroup = {
@@ -68,15 +68,28 @@ angular.module('main').controller('SingleGroupCtrl',['$scope', '$state', '$state
     GroupFactory.leaveGroup($stateParams.groupCode)
       .then(function() {
         GroupService.removeGroup($stateParams.groupCode);
+        $ionicHistory.nextViewOptions({
+            disableBack: true
+         });
         $state.go('tab.groups');
       })
       .catch($log.error);
   };
 
   $scope.endGroup = function(groupMembers) {
-    GroupFactory.endGroup(groupMembers, $stateParams.groupCode);
-    GroupService.removeGroup($stateParams.groupCode);
-    $state.go('tab.groups');
+    var confirmPopup = $ionicPopup.confirm({
+      title: 'End Group',
+      template: 'Are you sure? All your group info will be removed :(.'
+    });
+    confirmPopup.then(function(res) {
+      if (res) {
+        GroupFactory.endGroup(groupMembers, $stateParams.groupCode);
+        GroupService.removeGroup($stateParams.groupCode);
+        $ionicHistory.nextViewOptions({
+            disableBack: true
+         });
+        $state.go('tab.groups');
+      }
+    });
   };
-
 }]);
